Implemented judgement and note movement inside of the game loop
- Implemented movement of notes (in theory, it hasn't been tested yet)
- Implemented judgements according to rules I've stated (untested until next module is complete)
- Passed the hash set of keys held down inside of InputHistory so they can be used
- Removed some redundant code from the Judgement.cs file. JudgementEngineTests.cs has been updated to reflect updated structure
- Timestamping of inputs implemented inside of the game loop. This is working correctly.

+++ input/InputEvent.cs
@@ -5,6 +5,6 @@
     {
         public ushort VKey;
         public bool IsPressed;
-        public float TimeStamp;
+        public long TimeStamp;
     }
 }
